First assignment for ast121
Mentions dimensional analysis, in which we find planck length, time, and mass. It also includes a small mention of weak force interactions, with feynman diagrams. Then we are trying to estimate the mass of an arbitrary spiral galaxy, then ending with the density parameter and how it effects our universe.